requirements.txt
Flask 
requests 
python-dotenv
google-auth
celery
redis